[FSDP()][8/N] Refactor limiter's `_FreeEventQueue`
[ghstack-poisoned]
