+++ torch/distributed/fsdp/fully_sharded_data_parallel.py
@@ -1,4 +1,3 @@
-import collections
 import contextlib
 import copy
 import functools
@@ -13,7 +12,6 @@ from typing import (
     Any,
     Callable,
     cast,
-    Deque,
     Dict,
     Generator,
     Iterable,
@@ -58,6 +56,7 @@ from torch.distributed.fsdp._init_utils import (
     _move_module_to_device,
     _sync_module_states,
 )
+from torch.distributed.fsdp._limiter_utils import _FreeEventQueue
 from torch.distributed.fsdp._runtime_utils import (
     _clear_grads_if_needed,
     _prepare_forward_inputs,
@@ -588,35 +587,6 @@ class _ExecOrderData:
             self.current_order_index = 0
             if self.warn_status == _ExecOrderWarnStatus.WARNING:
                 self.warn_status = _ExecOrderWarnStatus.WARNED
-
-
-class _FreeEventQueue:
-    """
-    This tracks all pending frees corresponding to inflight all-gathers. The
-    queueing pattern is iterative enqueues with a single dequeue per iteration
-    once the limit ``_max_num_inflight_all_gathers`` is reached.
-    """
-
-    def __init__(self) -> None:
-        self._queue: Deque[torch.cuda.Event] = collections.deque()
-        self._max_num_inflight_all_gathers = 2  # empirically chosen
-
-    def enqueue(self, free_event: torch.cuda.Event) -> None:
-        """Enqueues a free event."""
-        self._queue.append(free_event)
-
-    def dequeue_if_needed(self) -> Optional[torch.cuda.Event]:
-        """Dequeues a single event if the limit is reached."""
-        if len(self._queue) >= self._max_num_inflight_all_gathers:
-            return self._dequeue()
-        return None
-
-    def _dequeue(self) -> Optional[torch.cuda.Event]:
-        """Dequeues a free event if possible."""
-        if self._queue:
-            event = self._queue.popleft()
-            return event
-        return None
 
 
 # TODO (awgu): Refactor this later
